Do not exclude `setuptools` from `pip freeze`
`pip freeze` has special logic to not output `distribute`, `setuptools`,
or `wheel` unless using Python 3.12 or higher. I created this repository
assuming the use of Python 3.12, so I did not encounter this problem
until using these files with Python 3.11. Because the code for older
Python versions is not ideal, I have set it to be conditional on the
version.

This code should be dropped in either 2025 Q4 (when the Scientific
Python ecosystem drops support for Python 3.11)[^1] or in 2027-10 (when
Python 3.11 reaches EOL).[^2]

[^1]: https://scientific-python.org/specs/spec-0000//#drop-schedule
[^2]: https://devguide.python.org/versions/